// src/features/companies/type.ts
export interface Company {
  company_name: string;
  description: string;
  mobile: string;
  logo_url: string | null;
  user_id: string;
  status: string;
  created_at: string;
  updated_at: string;
  email: string;
  username: string;
  company_id: string;
  role: string;
  user_status: string;
  user_created_at: string;
  user_updated_at: string;
  website?: string;
}